rework connect event
Currently the connect event fires everytime there's data if `!use.secure`. In theory this could be solved by using a `.once('connect')` listener, however this suffers from a delay during the initial connect if an environment is not using a username/password on the M1EXP.

When not using authentication, the current behavior is that no connect event is fired until some sort of data is received from the M1EXP. By default there is no connection banner or message sent by the M1EXP so even though the client is actually connected and ready, it must wait for the first transmission (either from an event or a regulary scheduled `XK` transmission from the m1exp, which could take up to 30 seconds.

This proposed change immedately emits the connect event after the tcp connection is established and reolaces the previous connect events with an `authorized` event when either the username/password is accepted or the first status message is output. It also fires these `authorized` event only once in the case of `!use.secure` by checking for and setting the `that._authorized` bool to true if not already set. Since in the case of `!use.secure` the client is essentially authorized.

A use case of eliminating the delay when no username/password is used:

```javascript
elk.on('connect', function (data) {
  elk.zoneDefinitionRequest()
}
```

Updated .gitignore